requirements.txt
wyoming>=1.6.0,<1.7.0
azure-cognitiveservices-speech>=1.41.1,<1.42
black>=24,<25
lxml>=5,<6
pycountry>=24.6.1,<25
ruff>=0.6.9,<0.7
pyaudio